refactor: use currencies context was replaced by a hook
- There was no need to use a context to keep the options of available
  and default currencies. The app is not too big to need it.

- useCurrenciesOptions hook was created.

- Condition was added to CurrencyExchangeCard to add the letter 'S' if
  to the label if the amount is bigger than 1.
